pyproject.toml
[project]
name = "xnano"
version = "0.0.44"
description = "build extremely 'nano' llm workflows"
readme = "readme.md"

requires-python = ">=3.9"

dependencies = [
    "bs4>=0.0.2",
    "duckduckgo-search>=6.3.4",
    "instructor>=1.6.3",
    "litellm>=1.52.3",
    "markdownify>=0.13.1",
    "pypdf2>=3.0.1",
    "qdrant-client>=1.12.1",
    "questionary>=2.0.1",
    "readability-lxml>=0.8.1",
    "rich>=13.9.4",
    "semchunk>=2.2.0",
    "sqlalchemy>=2.0.36",
]

[project.optional-dependencies]
fastembed = ["fastembed>=0.4.1"]
serve = ["nicegui", "fastapi", "uvicorn", "sse-starlette"]

[project.scripts] 
xnano = "xnano._cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mkdocs-material>=9.5.44",
    "mkdocs-material-extensions>=1.3.1",
    "mkdocstrings[python]>=0.27.0",
    "ruff>=0.7.3",
    "uv>=0.5.1",
    "mkdocs-jupyter>=0.25.1",
    "twine>=5.1.1",
]

[tool.hatch.build]
exclude = ["docs/"]
